Set the default operation name and resource name for web requests
This change sets the default 'operation.name' and 'resource.name' attributes for web requests.
The 'resource.name' is set to the concatenation of the HTTP method and the HTTP route.
The 'operation.name' is set to the string 'web.request'

The effect is that we can view all associated web requests in the UI by filtering on the 'operation.name' attribute.