feat(otel): add deployment ID resource attribute
# Why
Through log drain's `deploymentId` field it is possible to link log records to a specific Vercel deployment.

It would be great if the same level of linking could be established for OTLP data on Vercel. If that were the case, OTLP and log drain information could be correlated to the same "resource" (in OpenTelemetry terminology).

# What
Report `vercel.deployment_id` as a resource attribute.